feat: implemet error case

+++ server/src/modules/tasks/useCases/deleteTask/DeleteTaskUseCase.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from 'tsyringe'
+import { AppError } from '../../../../shared/errors/AppError'
 import { ITasksRepository } from '../../repositories/ITasksRepository'
 
 @injectable()
@@ -8,6 +9,12 @@ class DeleteTaskUseCase {
     private readonly tasksRepository: ITasksRepository) {}
 
   async execute (id: string): Promise<void> {
+    const deleteId = await this.tasksRepository.findTaskById(id)
+
+    if (deleteId === null) {
+      throw new AppError('Task do not exists!', 404)
+    }
+
     await this.tasksRepository.deleteTask(id)
   }
 }
